Added clustering by graph data
Added the option to cluster based on the graph data created from
assignments.csv. This option currently seems to work better than the
previous attempts, although further testing is neccessary. In the future,
a options page will need to be added, where the user will be able to
configure clustering settings.